requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0              # async PostgreSQL driver
psycopg2-binary==2.9.10      # sync driver for Alembic
alembic==1.14.0

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.0                # Google OAuth HTTP calls

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── File handling ─────────────────────────────────────────────────────────────
python-multipart==0.0.17     # FastAPI file uploads
boto3==1.35.80               # S3 storage
pypdf==5.1.0                 # PDF text extraction
python-docx==1.1.2           # DOCX text extraction

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.4.0            # optional JSON logging

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0                # async test client
factory-boy==3.3.1           # test data factories
